use version numbers for approved activity report display
